feat: improve resume and change download script to use md2pdf
