--- scripts/download-resume/markdown-to-html.ts
@@ -1,23 +1,0 @@
-export const markdownToHtml = (markdown: string): string => {
-  return markdown
-    .replace(/^### (.*$)/gm, '<h3>$1</h3>')
-    .replace(/^## (.*$)/gm, '<h2>$1</h2>')
-    .replace(/^# (.*$)/gm, '<h1>$1</h1>')
-    .replace(/^\*\*(.*?)\*\* \| (.*$)/gm, '<div class="job-title"><strong>$1</strong> | $2</div>')
-    .replace(/^\*(.*?)\*/gm, '<div class="job-period">$1</div>')
-    .replace(/^- \*\*(.*?)\*\* (.*$)/gm, '<li><strong>$1</strong> $2</li>')
-    .replace(/^- (.*$)/gm, '<li>$1</li>')
-    .replace(/^\*\*(.*?)\*\*/gm, '<strong>$1</strong>')
-    .replace(/^---$/gm, '<hr>')
-    .replace(/\n\n/g, '</p><p>')
-    .replace(/^(?!<)/gm, '<p>')
-    .replace(/(?<!>)$/gm, '</p>')
-    .replace(/<p><\/p>/g, '')
-    .replace(/<p>(<h[1-6]>)/g, '$1')
-    .replace(/(<\/h[1-6]>)<\/p>/g, '$1')
-    .replace(/<p>(<hr>)<\/p>/g, '$1')
-    .replace(/<p>(<div)/g, '$1')
-    .replace(/(<\/div>)<\/p>/g, '$1')
-    .replace(/<p>(<li>)/g, '<ul>$1')
-    .replace(/(<\/li>)<\/p>/g, '$1</ul>');
-};
