+++ src/pages/experience.mdx
@@ -2,7 +2,7 @@
 layout: '../layouts/Page.astro'
 title: 'My career journey'
 description: 'Julian Ramirez career journey'
-updatedDate: 2025-07-30
+updatedDate: 2025-08-08
 ---
 
 import Resume from "../components/sections/Resume.astro"
